requirements.txt
ollama
langchain-community
langchain-huggingface
faiss-cpu
fastmcp
pluggy >=1.5
platformdirs>=2
ipykernel
wikipedia
